==> pages/1_Identify_problem_and_data_clean.ipynb <==
import numpy as np         
import pandas as pd        
data = pd.read_csv('data/data.csv', index_col=False,)

# Id column is redundant and not useful, we want to drop it
data.drop('id', axis =1, inplace=True)
#data.drop('Unnamed: 0', axis=1, inplace=True)
data.head(2)

data.shape

data.info()

data.isnull().any()

data.diagnosis.unique()

data.to_csv('data/clean-data.csv')
